requirements.txt
Django>=5.0,<6.1
djangorestframework>=3.15,<3.17
Pillow>=10.0
whitenoise==6.12.0
python-dotenv==1.1.1
psycopg2-binary==2.9.12
gunicorn==25.3.0
dj-database-url==3.1.2
dotenv==0.9.9